+++ assimilation_code/modules/utilities/mpif08_utilities_mod.f90
@@ -1926,7 +1926,7 @@ subroutine get_from_mean(owner, window, mindex, x)
 
 integer,  intent(in)  :: owner  ! task in the window that owns the memory
 type(MPI_Win), intent(in)  :: window ! window object
-integer,  intent(in)  :: mindex ! index in the tasks memory
+integer(i8),   intent(in)  :: mindex ! index in the tasks memory
 real(r8), intent(out) :: x      ! result
 
 integer(KIND=MPI_ADDRESS_KIND) :: target_disp
@@ -1952,7 +1952,7 @@ subroutine get_from_fwd(owner, window, mindex, rows_in_window, num_rows, x)
 
 integer,  intent(in)  :: owner    ! task in the window that owns the memory
 type(MPI_Win),  intent(in)  :: window   ! window object
-integer,  intent(in)  :: mindex   ! index in the tasks memory
+integer(i8),    intent(in)  :: mindex   ! index in the tasks memory
 integer,  intent(in)  :: rows_in_window ! number of rows in the window
 integer,  intent(in)  :: num_rows ! number of rows to get from the window
 real(r8), intent(out) :: x(num_rows)     ! result
