+++ assimilation_code/modules/utilities/null_mpi_utilities_mod.f90
@@ -606,7 +606,7 @@ subroutine get_from_mean(owner, window, mindex, x)
 
 integer,  intent(in)  :: owner  ! task in the window that owns the memory
 integer,  intent(in)  :: window ! window object
-integer,  intent(in)  :: mindex ! index in the tasks memory
+integer(i8),  intent(in)  :: mindex ! index in the tasks memory
 real(r8), intent(out) :: x      ! result
 
 call error_handler(E_ERR,'get_from_mean', 'cannot be used in serial mode', source)
@@ -622,7 +622,7 @@ subroutine get_from_fwd(owner, window, mindex, rows_in_window, num_rows, x)
 
 integer,  intent(in)  :: owner    ! task in the window that owns the memory
 integer,  intent(in)  :: window   ! window object
-integer,  intent(in)  :: mindex   ! index in the tasks memory
+integer(i8),  intent(in)  :: mindex   ! index in the tasks memory
 integer,  intent(in)  :: rows_in_window ! number of rows in the window
 integer,  intent(in)  :: num_rows ! number of rows to get from the window
 real(r8), intent(out) :: x(num_rows)     ! result
