+++ models/POP/model_mod.f90
@@ -2139,7 +2139,7 @@ type(get_close_type),          intent(in)     :: gc
 type(location_type),           intent(inout)  :: base_loc, locs(:)
 integer,                       intent(in)     :: base_type, loc_qtys(:)
 integer(i8),                   intent(in)     :: loc_indx(:)
-integer,                       intent(out)    :: num_close, close_ind(:)
+integer(i8),                   intent(out)    :: num_close, close_ind(:)
 real(r8),            optional, intent(out)    :: dist(:)
 type(ensemble_type), optional, intent(in)     :: state_handle
