+++ models/clm/model_mod.f90
@@ -916,8 +916,8 @@ type(location_type),           intent(inout) :: locs(:)
 integer,                       intent(in)    :: base_type
 integer,                       intent(in)    :: loc_qtys(:)
 integer,                       intent(in)    :: loc_types(:)
-integer,                       intent(out)   :: num_close
-integer,                       intent(out)   :: close_ind(:)
+integer(i8),                   intent(out)   :: num_close
+integer(i8),                   intent(out)   :: close_ind(:)
 real(r8),            optional, intent(out)   :: dist(:)
 type(ensemble_type), optional, intent(in)    :: ens_handle
 
@@ -945,8 +945,8 @@ integer,                       intent(in)    :: base_type    !< observation TYPE
 type(location_type),           intent(inout) :: locs(:)      !< locations on my task
 integer,                       intent(in)    :: loc_qtys(:)  !< QTYs for locations on my task
 integer(i8),                   intent(in)    :: loc_indx(:)  !< indices into DART state on my task
-integer,                       intent(out)   :: num_close    !< how many are close
-integer,                       intent(out)   :: close_ind(:) !< indices into the locs array
+integer(i8),                   intent(out)   :: num_close    !< how many are close
+integer(i8),                   intent(out)   :: close_ind(:) !< indices into the locs array
 real(r8),            optional, intent(out)   :: dist(:)      !< distances (in radians)
 type(ensemble_type), optional, intent(in)    :: ens_handle
